fix: the same DA blobs for different Arbitrum batches (#11485)
* documentation for migration related modules

* New table introduced

* implementation of migration process

* proper data preparation for import batch-to-DA-blob records

* support of new DB schema in API

* spelling issues fixed

* format issues addressed

* credo issues addressed

* explicit references in the table arbitrum_batches_to_da_blobs

* added a comment about current limitation in API

* credo issue addressed

* env vars for configuration

* enable the migrator for :indexer only

* the list of env vars updated

* chore: split Arbitrum Reader and DB modules for improved maintainability (#11531)

* API related functionality in a separate module

* functions order adjusted

* functionality related to querying cross-chain messages info in a separate module

* functionality related to lifecycle transactions in a separate module

* changed locations of files

* separate module to keep queries related to settlement information

* get_anytrust_keyset moved to common reader functions

* feat: get all Arbitrum batches that corresponds to an AnyTrust data blob (#11541)

* initial support of getting all batches that correspond to AnyTrust DA blob

* credo issues fixed

* API Reader divided in several modules

* missed dependency added

* separate module for arbitrum-related events

* missed abi for unindexed params of the event

* separate module for arbitrum-related contracts

* code review comments addressed
