+++ apps/block_scout_web/lib/block_scout_web/controllers/api/v2/block_controller.ex
@@ -31,7 +31,7 @@ defmodule BlockScoutWeb.API.V2.BlockController do
   }
 
   alias Explorer.Chain
-  alias Explorer.Chain.Arbitrum.Reader, as: ArbitrumReader
+  alias Explorer.Chain.Arbitrum.Reader.API.Settlement, as: ArbitrumSettlementReader
   alias Explorer.Chain.Celo.ElectionReward, as: CeloElectionReward
   alias Explorer.Chain.Celo.EpochReward, as: CeloEpochReward
   alias Explorer.Chain.Celo.Reader, as: CeloReader
@@ -193,11 +193,10 @@ defmodule BlockScoutWeb.API.V2.BlockController do
       params
       |> select_block_type()
       |> Keyword.merge(paging_options(params))
-      |> Keyword.merge(@api_true)
 
     {blocks, next_page} =
       batch_number
-      |> ArbitrumReader.batch_blocks(full_options)
+      |> ArbitrumSettlementReader.batch_blocks(full_options)
       |> split_list_by_page()
 
     next_page_params = next_page |> next_page_params(blocks, delete_parameters_from_next_page_params(params))
