+++ docker-compose/envs/common-blockscout.env
@@ -274,6 +274,8 @@ INDEXER_DISABLE_INTERNAL_TRANSACTIONS_FETCHER=false
 # INDEXER_ARBITRUM_TRACKING_MESSAGES_ON_L1_RECHECK_INTERVAL=
 # INDEXER_ARBITRUM_MISSED_MESSAGES_RECHECK_INTERVAL=
 # INDEXER_ARBITRUM_MISSED_MESSAGES_BLOCKS_DEPTH=
+# ARBITRUM_DA_RECORDS_NORMALIZATION_MIGRATION_BATCH_SIZE=
+# ARBITRUM_DA_RECORDS_NORMALIZATION_MIGRATION_CONCURRENCY=
 # CELO_CORE_CONTRACTS=
 # INDEXER_CELO_VALIDATOR_GROUP_VOTES_BATCH_SIZE=200000
 # INDEXER_DISABLE_CELO_EPOCH_FETCHER=false
